Ensure containers for temporary data.
